feat(v3) Drop Notef, drop Cause(), drop errgo entirely
